+++ include/ParallelSDDPSolver.h
@@ -5,16 +5,11 @@
  * Header file for the ParallelSDDPSolver class, which derives from SDDPSolver
  * and implements a parallel version of SDDPSolver.
  *
- * \version 0.1
- *
- * \date 24 - 03 - 2021
- *
  * \author Rafael Durbano Lobato \n
- *         Operations Research Group \n
  *         Dipartimento di Informatica \n
  *         Universita' di Pisa \n
  *
- * \copyright &copy; by Rafael Durbano Lobato
+ * \copyright Copyright &copy; by Rafael Durbano Lobato
  */
 /*--------------------------------------------------------------------------*/
 /*----------------------------- DEFINITIONS --------------------------------*/
@@ -160,7 +155,7 @@ protected:
 
   Eigen::ArrayXd oneStepBackward
   ( const StOpt::SDDPCutOptBase & p_linCut,
-    const std::tuple< std::shared_ptr<Eigen::ArrayXd>, int, int > & p_aState,
+    const std::tuple< std::shared_ptr< Eigen::ArrayXd >, int, int > & p_aState,
     const Eigen::ArrayXd & p_particle, const int & p_isample) const override;
 
 /*--------------------------------------------------------------------------*/
@@ -200,7 +195,7 @@ protected:
   /** This function returns the index of a sub-Block of SDDPBlock at the given
    * \p stage whose data is set to the scenario whose index is \p
    * scenario_index. If no sub-Block at the given \p stage is set to the
-   * scenario specified, Inf<Index>() is returned.
+   * scenario specified, Inf< Index >() is returned.
    *
    * @param stage A stage between 0 and time_horizon - 1.
    *
@@ -208,7 +203,7 @@ protected:
    *
    * @return The index of a sub-Block of SDDPBlock at the given \p stage whose
    *         data is set to the scenario whose index is \p scenario_index. If
-   *         no such a sub-Block exists, Inf<Index>() is returned. */
+   *         no such a sub-Block exists, Inf< Index >() is returned. */
 
   Index get_sub_block_with_scenario( Index stage ,
                                      Index scenario_index ) const;