+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.2] - 2023-17-05
+
+### Added
+
+- Handle eEverykIteration events in SDDPSolver.
+
+### Fixed
+
+- Windows compilation issues.
+
 ## [0.5.1] - 2022-07-01
 
 ### Added
@@ -70,7 +80,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First test release.
 
-[Unreleased]: https://gitlab.com/smspp/sddpblock/-/compare/0.5.1...develop
+[Unreleased]: https://gitlab.com/smspp/sddpblock/-/compare/0.5.2...develop
+[0.5.2]: https://gitlab.com/smspp/sddpblock/-/compare/0.5.1...0.5.2
 [0.5.1]: https://gitlab.com/smspp/sddpblock/-/compare/0.5.0...0.5.1
 [0.5.0]: https://gitlab.com/smspp/sddpblock/-/compare/0.4.0...0.5.0
 [0.4.0]: https://gitlab.com/smspp/sddpblock/-/compare/0.3.0...0.4.0